// src/types/data.d.ts
export interface SmallMultipleRegion {
  region: 'East' | 'Central' | 'West' | 'NorthEast'
  n: number
  frailN: number
  frailRate: number
}